+++ src/protocol/dapio.cpp
@@ -154,6 +154,7 @@ void DAPIO::AddCapabilitiesTo(json &capabilities)
     capabilities["exceptionBreakpointFilters"] = excFilters;
     capabilities["supportsExceptionOptions"] = false; // TODO add implementation
     capabilities["supportsHitConditionalBreakpoints"] = true;
+    capabilities["supportsModulesRequest"] = true;
 }
 
 void DAPIO::SetupProtocolLogging(const std::string &path)