+++ test-suite/run_tests.ps1
@@ -39,6 +39,7 @@ $ALL_TEST_NAMES = @(
     "TestBreakpointWithoutStop"
     "TestUnhandledException"
     "TestStdIO"
+    "TestModules"
 )
 
 $TEST_NAMES = $tests